Move methods configuration parser to internal package
