+++ javaagent-tooling/src/main/java/io/opentelemetry/javaagent/tooling/config/internal/MethodsConfigurationParser.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.tooling.config;
+package io.opentelemetry.javaagent.tooling.config.internal;
 
 import static java.util.Collections.emptyList;
 import static java.util.Collections.emptyMap;
@@ -19,6 +19,10 @@ import java.util.Map;
 import java.util.Set;
 import java.util.logging.Logger;
 
+/**
+ * This class is internal and is hence not for public use. Its APIs are unstable and can change at
+ * any time.
+ */
 public final class MethodsConfigurationParser {
 
   private static final Logger logger = Logger.getLogger(MethodsConfigurationParser.class.getName());
